Crate tokio_ffmpeg_cli

Structs

Builder

A builder for an ffmpeg command

Encoder

ProgressEvent

An event about the encoding progress sent by ffmpeg

Enums

EncoderFromLineError

An error that occurs while parsing an Encoder from a line

Error

The error type

Event

An Event

LineBuilderError

An error that occurs while building a ProgressEvent

Functions

get_encoders

Get encoders that this ffmpeg supports