tokio_ffmpeg_cli

Struct Builder

pub struct Builder {
    pub audio_codec: Option<String>,
    pub video_codec: Option<String>,
    pub video_bitrate: Option<String>,
    pub input: Option<OsString>,
    pub output: Option<OsString>,
    pub input_format: Option<String>,
    pub output_format: Option<String>,
    pub pass: Option<u8>,
    pub video_frames: Option<u64>,
    pub video_profile: Option<String>,
    pub preset: Option<String>,
    pub overwrite: bool,
}

A builder for an ffmpeg command

Fields
audio_codec: Option<String>

The audio codec

video_codec: Option<String>

The video codec

video_bitrate: Option<String>

The video bitrate

input: Option<OsString>

The input

output: Option<OsString>

The output

input_format: Option<String>

The input format

output_format: Option<String>

The output format

pass: Option<u8>

The pass # for two pass

video_frames: Option<u64>

The # of video frames to read from the input

video_profile: Option<String>

The video profile

preset: Option<String>

The preset

overwrite: bool

Whether to overwrite the destination

Implementations

impl Builder

pub fn new() -> Self

Make a new Builder

pub fn audio_codec(&mut self, audio_codec: impl Into<String>) -> &mut Self

Set the audio codec

pub fn video_codec(&mut self, video_codec: impl Into<String>) -> &mut Self

Set the video codec

pub fn video_bitrate(&mut self, video_bitrate: impl Into<String>) -> &mut Self

Set the video bitrate

pub fn input(&mut self, input: impl Into<OsString>) -> &mut Self

Set the input

pub fn output(&mut self, output: impl Into<OsString>) -> &mut Self

Set the output

pub fn input_format(&mut self, input_format: impl Into<String>) -> &mut Self

Set the input format

pub fn output_format(&mut self, output_format: impl Into<String>) -> &mut Self

Set the output format

pub fn pass(&mut self, pass: u8) -> &mut Self

The pass # for 2 pass

pub fn video_frames(&mut self, video_frames: impl Into<u64>) -> &mut Self

The # of video frames to accept from the input

pub fn video_profile(&mut self, video_profile: impl Into<String>) -> &mut Self

The profile of the video

pub fn preset(&mut self, preset: impl Into<String>) -> &mut Self

The preset

pub fn overwrite(&mut self, overwrite: bool) -> &mut Self

Set whether the output should be overwritten

pub async fn ffmpeg_status(&mut self) -> Result<ExitStatus, Error>

Run the command and wait for it to finish.

This will not provide progress info or stdout/stdin, but is far simpler to drive.

pub async fn ffmpeg_output(&mut self) -> Result<Output, Error>

Run the command and wait for it to finish.

This will not provide progress info, but is far simpler to drive.

pub fn spawn( &mut self, ) -> Result<impl Stream<Item = Result<Event, Error>> + Unpin, Error>

Spawn the stream